Move sd notification to the right place and factor out functionality into module
