+++ jay-config/src/_private/client.rs
@@ -46,21 +46,11 @@ use {
     std::{
         cell::{Cell, RefCell},
         collections::{HashMap, VecDeque, hash_map::Entry},
-        env,
-        ffi::OsStr,
         future::Future,
-        io, mem,
+        mem,
         ops::Deref,
-        os::{
-            fd::IntoRawFd,
-            linux::net::SocketAddrExt,
-            unix::{
-                ffi::OsStrExt,
-                net::{SocketAddr, UnixDatagram},
-            },
-        },
+        os::fd::IntoRawFd,
         panic::{AssertUnwindSafe, catch_unwind},
-        path::Path,
         pin::Pin,
         ptr,
         rc::Rc,
@@ -85,30 +75,6 @@ fn run_cb<T>(name: &str, cb: &Callback<T>, t: T) {
         Ok(mut cb) => ignore_panic(name, || cb(t)),
         Err(_) => log::error!("Cannot invoke {name} callback because it is already running"),
     }
-}
-
-fn send_sd_notify_if_enabled(msg: &[u8]) {
-    match env::var_os("NOTIFY_SOCKET") {
-        Some(notify_socket) => {
-            if let Err(e) = try_send_sd_notify(msg, notify_socket.as_os_str()) {
-                log::error!("Failed to send systemd ready notification: {e}");
-            }
-        }
-        None => {
-            log::debug!("Not sending sd notification, NOTIFY_SOCKET not set");
-        }
-    }
-}
-
-fn try_send_sd_notify(msg: &[u8], path: &OsStr) -> io::Result<()> {
-    let addr = match path.as_bytes().strip_prefix(b"@") {
-        Some(abs) => SocketAddr::from_abstract_name(abs)?,
-        None => SocketAddr::from_pathname(Path::new(&path))?,
-    };
-    let sock = UnixDatagram::unbound()?;
-    sock.send_to_addr(msg, &addr)?;
-
-    Ok(())
 }
 
 fn ignore_panic(name: &str, f: impl FnOnce()) {
@@ -2366,7 +2332,6 @@ impl ConfigClient {
                 }
             }
             ServerMessage::GraphicsInitialized => {
-                send_sd_notify_if_enabled(b"READY=1");
                 if let Some(handler) = self.on_graphics_initialized.take() {
                     ignore_panic("graphics initialized", handler);
                 }
