+++ src/utils.rs
@@ -61,6 +61,7 @@ pub mod refcounted;
 pub mod reset;
 pub mod run_toplevel;
 pub mod scroller;
+pub mod sd_notify;
 pub mod send_sync_rc;
 pub mod smallmap;
 pub mod stack;
